changed names for register-files and added admin=0
