+++ includes/register_controller.php
@@ -2,14 +2,14 @@
 session_start();
 
 require_once ("database_connection.php");
-require_once ("../classes/Register.php");
+require_once ("../classes/RegisterMethods.php");
 
 /* register_controller is the "central" for Register and here is were al the functions in Register are initialized and the instances of both classes RegisterDataSetup and Register are initalized. Register_controller also contains error-handling and $_POST Globals */
 
 $register_username = $_POST['username'];
 $register_password = $_POST['password']; 
 
-$registration_steps = new Register($pdo);
+$registration_steps = new RegisterMethods($pdo);
 /* ***Error handling */
 
 //If the provided password is shorter than 6 chars - display error.
@@ -20,7 +20,7 @@ if (strlen($register_password) <= 6)
 }
 
 //Create an instance with the data from the form
-$new_user = new RegisterDataSetup($register_username, $register_password);
+$new_user = new Register($register_username, $register_password);
 //Check if register_username is avalible with the new instance of user
 
 $row = $registration_steps->check_for_user_in_database($new_user);
